value-pairs: Type hinting support
This implements very basic support for type hinting template strings,
where those hints will be available to value-pairs too, allowing any
user of either LogTemplate or value-pairs to handle different types of
values than plain strings. How they handle it, is up to the callers,
this patch just lays the groundwork that makes it possible to pass
these hints around.

Furthermore, this includes support for the on-error() statement, which
tells drivers how type-casting failures shall be handled: they can
either drop the message, drop the property, or fall back to string;
and can do either of these either verbosely, logging the fact, or
silently.

This fixes #4 and #5.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>